Basic project structure.
Django url routers, basic models and configuration have been added.
Note the empty User model. Since Django heavily depends on the User
class, it can't be modified after the first migration. So, it has
been subclassed in case modifications are required.

Scripts to start and stop the server are included.

A standard configurator has been included to change settings between
dev and production, including details for nginx and django itself.
Specific parameters can be found using python configure.py --help.

Note that for dev, SQLite will be used, whilst MySQL is used in
production.

The configurator will generate dummy tokens for Django and MySQL
in the ltm/token folder. The Django one is enough for real use
(since it's only a random string), but the MySQL one should be
edited to reflect the correct credentials used to access the database.