Basic UI.
The main UI components have been added.

The page is composed of a header, a main section and a footer.

Inside the main section, the website shows the details for the
current user and a list of all the users.

If no user has been selected, a landing message is displayed.
If the username is invalid, an error message is shown.

No styling has been applied yet.