Added Search.
The users can now search for people on the directory through a
simple search box.

This box does a straightforward block search, i.e. the entire text
is used as the target -> "Doe John" will not find "John Doe".
The term will be search in the name, the email, the phone and
the address. The search is case insensitive as it's natural for
human input.