Styling corrections.
Addresses now convert newlines into <br/> tags so the format looks
right.

The scrollbar in the sidebar is hidden if not necessary. Also, its
width is now correctly enforced (it was previously overriden by
the other cell width, which takes 100%, so the sidebar got squeezed
to its minimum acceptable width). This was done by using min-width
instead of width.