Truncate user list.
The sidebar userlist now truncates at 100 users so it doesn't bog
the website down. If the user can't see the person she wants to
see, they should search for them.

Naturally, this could be extended a system of "pages" or infinite
scrolling.

At this point, the website is completely usable.