Responsive design improvements.
The main UI used a table display, which does not play well with mobile
devices. A media query now changes this into a flexbox if the width
is too small.

The flexbox is only used to change the order of the elements (so
the search is below the user details). It may need to be reviewed
for old browser compatibility.