Added unit tests.
Testing for all the server business logic has been added. This includes
the user model, the userlist functions and the userlist view.

With the help of these tests some bugs have been fixed, such as
no-name users and trailing spaces in full names.

Validation has been included for the search box.

Still missing: validation for the username url and tests for the
loadusers command.