Client app.
The client site has been augmented with React so the experience is
more fluid. The site no longer needs to fetch the entire site to
change a small component.

The user can load the user list in chunks asynchronously.

The new client app needs testing.

If the client is not modern enough to understand React, the basic
static version is still there (React modifies it after the fact).