Client app authorization headers.
The client app was not forwarding the authorization headers, so if
the server was behind a basic-http-auth, the app would fail to
be interactive, since every JSON call would return "forbidden".

The app now sends the appropriate authorization headers.

Also, some validations were corrected and the style of parts
of the code has been improved.